# docs/openapi.json
{
    "openapi": "3.0.1",
    "info": {
        "title": "GlazeAudio API",
        "description": "REST API for GlazeAudio \u2013 a platform for reviewing music. Resources are hierarchical: albums \u2192 songs \u2192 reviews. Reviews are written by users.",
        "version": "v1"
    },
    "servers": [
        {
            "url": "http://localhost:5080/"
        }
    ],
    "paths": {
...
        "/api/albums": {
            "get": {
                "tags": [
                    "Albums"
                ],
                "summary": "List albums (paged, filterable)",
                "description": "Returns one page of albums with their song count, review count and average rating. Filter with search, artist, genre, fromYear and toYear. Paging metadata and first/prev/next/last links are included in the response.",
                "operationId": "GetAlbums",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Page number, starting at 1. Default 1.",
                        "schema": {
                            "type": "integer",
                            "format": "int32"
                        }
                    },
                    {
                        "name": "pageSize",
                        "in": "query",
                        "description": "Items per page, 1\u201350. Default 10.",
                        "schema": {
                            "type": "integer",
                            "format": "int32"
                        }
                    },
                    {
                        "name": "search",
                        "in": "query",
                        "description": "Text to find in the album title or artist (case-insensitive).",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "artist",
                        "in": "query",
                        "description": "Only albums whose artist contains this text.",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "genre",
                        "in": "query",
                        "description": "Only albums whose genre contains this text, e.g. 'rock'.",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "fromYear",
                        "in": "query",
                        "description": "Only albums released in or after this year.",
                        "schema": {
                            "type": "integer",
                            "format": "int32"
                        }
                    },
                    {
                        "name": "toYear",
                        "in": "query",
                        "description": "Only albums released in or before this year.",
                        "schema": {
                            "type": "integer",
                            "format": "int32"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PagedResultOfAlbumDto"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "application/problem+json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HttpValidationProblemDetails"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Albums"
                ],
                "summary": "Create an album",
                "description": "Creates a new album. Returns 201 with a Location header that points to the new album.",
                "operationId": "CreateAlbum",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/AlbumRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "application/problem+json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProblemDetails"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "content": {
                            "application/problem+json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HttpValidationProblemDetails"
                                }
                            }
                        }
                    },
                    "201": {
                        "description": "Created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/AlbumDto"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/albums/{albumId}": {
            "get": {
                "tags": [
                    "Albums"
                ],
                "summary": "Get one album",
                "operationId": "GetAlbum",
                "parameters": [
                    {
                        "name": "albumId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int32"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/AlbumDto"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "content": {
                            "application/problem+json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProblemDetails"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Albums"
                ],
                "summary": "Update an album",
                "description": "Replaces all editable fields of the album.",
                "operationId": "UpdateAlbum",
                "parameters": [
                    {
                        "name": "albumId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int32"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/AlbumRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "application/problem+json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProblemDetails"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "content": {
                            "application/problem+json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HttpValidationProblemDetails"
                                }
                            }
                        }
                    },
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/AlbumDto"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "content": {
                            "application/problem+json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProblemDetails"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Albums"
                ],
                "summary": "Delete an album",
                "description": "Deletes the album together with all of its songs and their reviews. Returns 204 with no body.",
                "operationId": "DeleteAlbum",
                "parameters": [
                    {
                        "name": "albumId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int32"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "404": {
                        "description": "Not Found",
                        "content": {
                            "application/problem+json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProblemDetails"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/albums/{albumId}/overview": {
            "get": {
                "tags": [
                    "Albums"
                ],
                "summary": "Album overview (album + songs + reviews)",
                "description": "A dashboard-style resource built from three entities: the album, its songs with their ratings, rating statistics per aspect across all reviews, the top-rated song and the latest reviews.",
                "operationId": "GetAlbumOverview",
                "parameters": [
                    {
                        "name": "albumId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int32"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/AlbumOverviewDto"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "content": {
                            "application/problem+json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProblemDetails"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "AlbumDto": {
                "required": [
                    "id",
                    "title",
                    "artist",
                    "releaseYear",
                    "genre",
                    "coverUrl",
                    "songCount",
                    "reviewCount",
                    "averageRating"
                ],
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "title": {
                        "type": "string"
                    },
                    "artist": {
                        "type": "string"
                    },
                    "releaseYear": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "genre": {
                        "type": "string"
                    },
                    "coverUrl": {
                        "type": "string",
                        "nullable": true
                    },
                    "songCount": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "reviewCount": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "averageRating": {
                        "type": "number",
                        "format": "double",
                        "nullable": true
                    },
                    "_links": {
                        "type": "object",
                        "additionalProperties": {
                            "$ref": "#/components/schemas/Link"
                        }
                    }
                }
            },
            "AlbumOverviewDto": {
                "required": [
                    "album",
                    "stats",
                    "songs",
                    "topRatedSong",
                    "latestReviews"
                ],
                "type": "object",
                "properties": {
                    "album": {
                        "$ref": "#/components/schemas/AlbumSummary"
                    },
                    "stats": {
                        "$ref": "#/components/schemas/AlbumStats"
                    },
                    "songs": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/OverviewSong"
                        }
                    },
                    "topRatedSong": {
                        "$ref": "#/components/schemas/OverviewSong2"
                    },
                    "latestReviews": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/OverviewReview"
                        }
                    },
                    "_links": {
                        "type": "object",
                        "additionalProperties": {}
                    }
                }
            },
            "AlbumRequest": {
                "required": [
                    "title",
                    "artist",
                    "releaseYear",
                    "genre",
                    "coverUrl"
                ],
                "type": "object",
                "properties": {
                    "title": {
                        "maxLength": 200,
                        "minLength": 1,
                        "type": "string",
                        "nullable": true
                    },
                    "artist": {
                        "maxLength": 200,
                        "minLength": 1,
                        "type": "string",
                        "nullable": true
                    },
                    "releaseYear": {
                        "maximum": 2100,
                        "minimum": 1900,
                        "type": "integer",
                        "format": "int32",
                        "nullable": true
                    },
                    "genre": {
                        "maxLength": 100,
                        "minLength": 1,
                        "type": "string",
                        "nullable": true
                    },
                    "coverUrl": {
                        "maxLength": 500,
                        "minLength": 0,
                        "type": "string",
                        "format": "uri",
                        "nullable": true
                    }
                }
            },
            "AlbumStats": {
                "required": [
                    "songCount",
                    "reviewCount",
                    "totalDurationSeconds",
                    "averageRating",
                    "aspectAverages"
                ],
                "type": "object",
                "properties": {
                    "songCount": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "reviewCount": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "totalDurationSeconds": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "averageRating": {
                        "type": "number",
                        "format": "double",
                        "nullable": true
                    },
                    "aspectAverages": {
                        "$ref": "#/components/schemas/AspectAverages"
                    }
                }
            },
            "AlbumSummary": {
                "required": [
                    "id",
                    "title",
                    "artist",
                    "releaseYear",
                    "genre",
                    "coverUrl"
                ],
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "title": {
                        "type": "string"
                    },
                    "artist": {
                        "type": "string"
                    },
                    "releaseYear": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "genre": {
                        "type": "string"
                    },
                    "coverUrl": {
                        "type": "string",
                        "nullable": true
                    }
                }
            },
...
            "HttpValidationProblemDetails": {
                "type": "object",
                "properties": {
                    "type": {
                        "type": "string",
                        "nullable": true
                    },
                    "title": {
                        "type": "string",
                        "nullable": true
                    },
                    "status": {
                        "type": "integer",
                        "format": "int32",
                        "nullable": true
                    },
                    "detail": {
                        "type": "string",
                        "nullable": true
                    },
                    "instance": {
                        "type": "string",
                        "nullable": true
                    },
                    "errors": {
                        "type": "object",
                        "additionalProperties": {
                            "type": "array",
                            "items": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "Link": {
                "required": [
                    "href"
                ],
                "type": "object",
                "properties": {
                    "href": {
                        "type": "string"
                    },
                    "method": {
                        "type": "string",
                        "default": "GET"
                    }
                }
            },
...
            "OverviewSong": {
                "required": [
                    "id",
                    "trackNumber",
                    "title",
                    "durationSeconds",
                    "reviewCount",
                    "averageRating"
                ],
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "trackNumber": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "title": {
                        "type": "string"
                    },
                    "durationSeconds": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "reviewCount": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "averageRating": {
                        "type": "number",
                        "format": "double",
                        "nullable": true
                    },
                    "_links": {
                        "type": "object",
                        "additionalProperties": {
                            "$ref": "#/components/schemas/Link"
                        }
                    }
                }
            },
...
            "PagedResultOfAlbumDto": {
                "required": [
                    "items",
                    "page",
                    "pageSize",
                    "totalCount",
                    "totalPages"
                ],
                "type": "object",
                "properties": {
                    "items": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/AlbumDto"
                        }
                    },
                    "page": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "pageSize": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "totalCount": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "totalPages": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "_links": {
                        "type": "object",
                        "additionalProperties": {}
                    }
                }
            },
...
            "ProblemDetails": {
                "type": "object",
                "properties": {
                    "type": {
                        "type": "string",
                        "nullable": true
                    },
                    "title": {
                        "type": "string",
                        "nullable": true
                    },
                    "status": {
                        "type": "integer",
                        "format": "int32",
                        "nullable": true
                    },
                    "detail": {
                        "type": "string",
                        "nullable": true
                    },
                    "instance": {
                        "type": "string",
                        "nullable": true
                    }
                }
            },
...
        }
    },
...
}